Removed env variables from backend that aren't necessary to deploy on Vercel. Added a env to client.
