+++ backend/src/config.ts
@@ -7,8 +7,6 @@ import dotenv from 'dotenv';
 dotenv.config({ path: "./config.env" });
 
 const config = {
-    port: 5000,
-    host: process.env.HOST || 'localhost',
     mongoURI: process.env.MONGO_URI,
     apiKey: process.env.API_KEY,
 }
