+++ client/src/config.env
@@ -1,0 +1,1 @@
+VITE_API_BASE_URL =
